+++ src/lib/pages/account-details/components/delegations/radio-card/index.tsx
@@ -1,9 +1,6 @@
 import { Radio } from "@chakra-ui/react";
-import type { Big } from "big.js";
-import big from "big.js";
 
-import type { DenomInfo } from "lib/pages/account-details/types";
-import type { Option, TokenWithValue, USD } from "lib/types";
+import type { Option, TokenWithValue } from "lib/types";
 
 import { MultiBondsRadioCard } from "./MultiBondsRadioCard";
 import { SingleBondRadioCard } from "./SingleBondRadioCard";
@@ -12,7 +9,7 @@ interface RadioCardProps {
   value: string;
   tokens: Option<Record<string, TokenWithValue>>;
   isLoading: boolean;
-  bondDenoms: DenomInfo[];
+  bondDenoms: TokenWithValue[];
 }
 
 export const RadioCard = ({
@@ -26,13 +23,7 @@ export const RadioCard = ({
       <SingleBondRadioCard
         value={value}
         token={
-          tokens
-            ? Object.values(tokens)[0] ?? {
-                ...bondDenoms[0],
-                amount: 0,
-                value: big(0) as USD<Big>,
-              }
-            : undefined
+          tokens ? tokens[bondDenoms[0].denom] ?? bondDenoms[0] : undefined
         }
         isLoading={isLoading}
       />